NUTCH-2805: Rename plugin urlfilter-domainblacklist (#540)
NUTCH-2805: Rename plugin urlfilter-domainblacklist
