+++ conf/domaindenylist-urlfilter.txt.template
@@ -13,4 +13,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-# config file for urlfilter-domainblacklist plugin
+# config file for urlfilter-domaindenylist plugin
