+++ default.properties
@@ -103,7 +103,7 @@ plugins.protocol=\
 plugins.urlfilter=\
    org.apache.nutch.urlfilter.automaton*:\
    org.apache.nutch.urlfilter.domain*:\
-   org.apache.nutch.urlfilter.domainblacklist*:\
+   org.apache.nutch.urlfilter.domaindenylist*:\
    org.apache.nutch.urlfilter.fast*:\
    org.apache.nutch.urlfilter.ignoreexempt*:\
    org.apache.nutch.urlfilter.prefix*:\
