+++ src/plugin/urlfilter-domaindenylist/build.xml
@@ -15,7 +15,7 @@
  See the License for the specific language governing permissions and
  limitations under the License.
 -->
-<project name="urlfilter-domainblacklist" default="jar-core">
+<project name="urlfilter-domaindenylist" default="jar-core">
 
   <import file="../build-plugin.xml"/>
